Refactor Code View into dedicated CodeView component
Moved all Monaco editor logic, toolbar, and error handling from WorkflowDefinitionWorkspace into a new CodeView component. CodeView now manages JSON editing, auto-apply/manual apply, error display, and throttling. Updated parent workspace to delegate code view responsibilities to CodeView and simplified its logic. Improved separation of concerns and maintainability. Also fixed UI details and moved related CSS.
