+++ src/modules/Elsa.Studio.Workflows/Components/WorkflowDefinitionEditor/Components/WorkflowDefinitionWorkspace.razor
@@ -28,17 +28,6 @@
         flex-direction: column;
         min-height: 0;
     }
-
-    .editor-host {
-        flex: 1 1 auto;
-        min-height: 0;
-        overflow: hidden;
-    }
-
-    .code-view-monaco-editor {
-        height: 100% !important;
-        width: 100% !important;
-    }
 </style>
 
 <CascadingValue Value="this">
@@ -82,43 +71,10 @@
                                  ShowCloseIcon="false"
                                  Style="height:100%;">
                         <div class="code-tab-content">
-                            <MudPaper Class="ma-2" Elevation="0" Outlined="true">
-                                <MudToolBar Dense="true">
-                                    <MudSpacer />
-                                    <MudTooltip Text="@Localizer["Auto apply changes to editor."]">
-                                        <MudSwitch T="bool"
-                                                   @bind-Value="@_autoApply"
-                                                   Color="Color.Primary"
-                                                   Label="@Localizer["Auto-apply"]" />
-                                    </MudTooltip>
-                                    <MudTooltip Text="@Localizer["Apply changes to editor."]">
-                                        <MudIconButton Icon="@Icons.Material.Outlined.Check"
-                                                       Color="Color.Success"
-                                                       Label="@Localizer["Apply"]"
-                                                       OnClick="OnApplyClicked" />
-                                    </MudTooltip>
-                                </MudToolBar>
-                                @if (!string.IsNullOrWhiteSpace(_applyErrorMessage))
-                                {
-                                    <MudAlert Severity="Severity.Error" Dense="true">
-                                        <MudStack Row="true" AlignItems="MudBlazor.AlignItems.Center">
-                                            @_applyErrorMessage
-                                            <MudTooltip Text="@Localizer["Reload from editor"]">
-                                                <MudIconButton Icon="@Icons.Material.Outlined.Refresh"
-                                                               Size="Size.Small"
-                                                               OnClick="ReloadMonacoClick" />
-                                            </MudTooltip>
-                                        </MudStack>
-                                    </MudAlert>
-                                }
-                            </MudPaper>
-                            <div class="editor-host">
-                                <StandaloneCodeEditor @ref="_monacoEditor"
-                                                      Id="@_monacoEditorId"
-                                                      ConstructionOptions="ConfigureMonacoEditor"
-                                                      OnDidChangeModelContent="OnMonacoContentChangedAsync"
-                                                      CssClass="code-view-monaco-editor" />
-                            </div>
+                            <CodeView @ref="CodeView"
+                                      @bind-AutoApply="_autoApply"
+                                      WorkflowDefinitionSerialized="@WorkflowDefinitionSerialized"
+                                      ApplyWorkflowDefinition="OnChildApply" />
                         </div>
                     </MudTabPanel>
                 </MudDynamicTabs>
